requirements.txt
asgiref==3.2.7
Django==3.0.5
django-cors-headers==3.2.1
django-rest-framework==0.1.0
djangorestframework==3.11.0
gunicorn==20.0.4
joblib==0.14.1
numpy==1.18.3
pandas==1.0.3
python-dateutil==2.8.1
pytz==2019.3
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
sqlparse==0.3.1
whitenoise==5.0.1
